add_data_module/requirements.txt
Flask==2.1.1
matplotlib==3.4.3
mysql-connector-python==8.0.27
numpy==1.21.2
plotly==5.3.1
requests==2.31.0